requirements.txt
pandas
numpy
scikit-learn
joblib
mlflow
huggingface_hub
python-dotenv
streamlit
pyyaml
xgboost
